refactor: extract minimal shared logic between HTML and EPUB formatters
