+++ test/ex_doc/formatter/epub/templates_test.exs
@@ -1,7 +1,7 @@
 defmodule ExDoc.Formatter.EPUB.TemplatesTest do
   use ExUnit.Case, async: true
 
-  alias ExDoc.Formatter.HTML
+  alias ExDoc.{Formatter}
   alias ExDoc.Formatter.EPUB.Templates
 
   defp source_url do
@@ -28,7 +28,7 @@ defmodule ExDoc.Formatter.EPUB.TemplatesTest do
   defp get_module_page(names, config \\ []) do
     config = doc_config(config)
     {mods, []} = ExDoc.Retriever.docs_from_modules(names, config)
-    [mod | _] = HTML.render_all(mods, [], ".xhtml", config, highlight_tag: "samp")
+    [mod | _] = Formatter.render_all(mods, [], ".xhtml", config, highlight_tag: "samp")
     Templates.module_page(config, mod)
   end
